requirements.txt
httpx
beautifulsoup4
lxml
requests
anthropic
google-genai
python-dotenv
